requirements.txt
pylint==2.9.3
pytest
click
black==20.8b1
pytest-cov
requests==2.25.1
pymongo==3.11.2
dnspython==2.0.0
flask==1.1.2
numpy
pandas
bs4
flask
PyYAML>=5.4